Fix the build on Mac OS X 10.5.
* Find Python using the appropriate CMake commands (will find python 2.6 from MacPorts instead of python 2.5 which ships with Leopard and 2.6 is required)
* Shut up compiler warnings about variables that might be used uninitialized
* Shut up one unsigned comparison < 0 is always false warning in test/m88k
* Loosen the -Werror requirement for libnix (flex generated code emits a warning) and m88k (Untested on this platform warning shouldn't make the entire build fail)
* Cast to force the argument to ntohl to be 32 bits.

+++ test/m68k/amigaos_lib.cpp
@@ -148,7 +148,7 @@
 #define CAST_ADDR(a) ((a)&0xFFFF) /* HACK: limit address bus to 16 bit */
 #define CAST32(a) (*(uint32_t*)(&(a)))
 #define RAM32(a) (CAST32(RAM[CAST_ADDR(a)]))
-#define STORE32(a,b) (RAM32(a) = htonl(b)) 
+#define STORE32(a,b) (RAM32(a) = htonl((uint32_t)(b))) 
 uint8_t *RAM;
 
 void
